Created float parser for laundry data
This test data should allow to work with the coming recommendation
engine in order to begin to make predictions about the current state
of the clothing needs. Due to last minute changes in the code, sensor
currently strings.Replace removes more things then neccesary, consider
adding this in a todo.